Add method to brewery service to retrieve all beers for a brewery
Adds method and test for same to retrieve all beers
associated with a specific brewery. All tests are
currently passing. Output of method is an array of
hashes as expected. Separated responsibility for this
out to a separate method for purpose of SRP, and I think
it will be less clunky overall to drive and manipulate the
data.