Complete brewery index styling
Adds styling to brewery index. Each brewery is an image
card with the brewery image, name, and a link to their
website. The description was removed from this page to
prevent redundancy between index and show as well as
due to severe inconsistencies between the length/presence
of a description for each brewery. Test was modified to
accomodate this change. All tests currently passing.
Test coverage at 99.08%