Add styling for brews on brewery show page
Adds cards similar to brewery index for each instance of a beer
returned that belongs to a brewery. Card size remains an issue
due to variance in description length. Discovered bug: if a
brewery does not have any beers listed in the DB there will
be a nil class error in the Brew model all_at_brewery method.
Will need to circumvent this.