Add function for navigating pages of brewery index
Adds two buttons to brewery index: forward and backward.
Same work with a helper method(current_page) to determine
the current page number and increment it up or down in the
parameter of brewery_page passed to the controller from the
buttons. Adds helper method(check_page?) which determines
if the number is less than one; if this is the case, the
controller will reset the parameter to one and render the
view.