Add search features
Adds search controller with show method. Show page
for search includes a jumbotron with search forms
for events and breweries by city. A link has been
added to the navigation bar to link to the search
page for authenticated users. Intention is to utse
AJAX queries to on that page for displaying search
results.