Modify styling of brewery show
Brews will be displayed in a single column as
breweries are on brewery index. Reformatted
#brewery-location-contact to be the same
length as the #brewery-show-row and aligned
the linkt to the website to the right of this div.
Adds a banner above #brew_list with an h4
identifying the below contents as the brews for
that brewery. All tests remain passing.