Add authorize action
Adds authorize! method to application controller.
Same will redirect a user to root unless current_user
is accounted for. A before_action was set in the
breweries and search controllers with same. Feature
tests modified to stub in a current user. Sad path
testing for guests drafted and passing. All tests
currently passing with test coverage at 98.75%. Also
changes test in search card at /search to be white in
color.