Add user favorites function
Adds feature for the user to be able to add a
brewery as a favorite/place to go. ActiveRecord
models added for FavoriteBrewery and UsersFavoriteBrewery.
Favorite brewery has only uid as an attribute(the id
used by the API to identify it. UsersFavoriteBrewery
is a joins between users and favorite_breweries.
Feature test confirming functionality of button is
currently passing. The action lives in the show method
of the FavoriteBreweries controller. All tests
currently passing.