Lay framework for user favorite endpoint
Adds api/v1 endpoint for the purpose of accessing an
individual brewery. Intention is to have iterative
AJAX queries based on favorites of current_user.
Adds api/v1/favorite_breweries_controller with show
method. Same calls to a new method living in the
brewery model which responds with a hash of brewery
data based on a brewery id being passed through.
Test added to brewery spec to see that new
#get_user_favorite method is functioning as
intended. Adds request spec test to see that the
new API endpoint is functioning as intended. All
tests are currently passing with coverage at 98.92%.
Also adds header to user show page to act as label
for their favorites. Adds empty div #user-favorites
to bottom of show page to be targeted for placement
of brewery cards via JS.